Updated room generator
* The room generator and chamber generator are so similar, except for
the table name and the exit generator.  I should make a single
generator, pass in the table name and exit generator as dependencies,
and that's that.  But if the logic were to diverge, it would be handy to
still have them as separate things.
* I hope to have some time at work today to get some more done, and
hopefully I will be able to find the random door type for a random
dungeon table online